Rename fields, add endpoints for route generation & points list
I've decided to roll out my own graph code, as:
  - graph-tool.skewed.de - complicated build process, can't just download from pip
  - igraph - you cannot catch exceptions from used C++ code, as they just kill the process
  - NetworkX - slow?

It's suboptimal, but since we're doing a MVP, it's a waste of time to optimize prematurely.
We can improve this if either the server gets overwhelmed,
  or if we plan on deploying on another campus.
The graph.py file provides a layer of abstraction, so it should be easy to replace.